+++ crates/compiler/src/mml/tokenizer.rs
@@ -45,6 +45,7 @@ pub enum Token<'a> {
     Wait,
     PlayPitch,
     PlayPitchSampleRate,
+    PlayPitchFrequency,
     PlayNoise,
     DisableNoise,
     SetOctave,
@@ -428,6 +429,7 @@ fn next_token<'a>(scanner: &mut Scanner<'a>) -> Option<TokenWithPosition<'a>> {
 
         b'P' => match scanner.second_byte() {
             Some(b'M') => two_ascii_token!(Token::PitchMod),
+            Some(b'F') => two_ascii_token!(Token::PlayPitchFrequency),
             Some(b'R') => two_ascii_token!(Token::PlayPitchSampleRate),
             _ => one_ascii_token!(Token::PlayPitch),
         },